refactor and re organized the app compoents in the herarchy
